Rename testbench.rs -> test_bench.rs
